Fix 3 bugs found while verifying the merge against a fresh staging database
- bootstrap.js: the auto-created default workspace wasn't linked to the org
  on a truly fresh database (no existing workspace) — every org-scoped query
  (editors, tasks, time rules) silently returned nothing. Production never
  hit this because a workspace already existed before the org concept was
  introduced; a clean bootstrap does.
- tasks.js (POST /tasks, spawnNextOccurrence): "could not determine data
  type of parameter" — a numeric parameter reused inside a bare `is not
  null` CASE check needs an explicit ::numeric cast for Postgres to infer
  its type when it's NULL.
- tasks.tsx: accepting a task showed it as "scheduled for later" for up to
  60s (the countdown only re-ticks once a minute) — bump the clock
  immediately after accepting so it reflects the real elapsed time right away.

Verified against a live staging Supabase project (all 27 migrations
applied): login, dashboard, task creation with type/format badges,
assign-to-someone-else pending state, accept flow, and the office-hours
dual-confirmation all confirmed working end-to-end.

+++ server/src/routes/tasks.js
@@ -140,7 +140,7 @@ tasksRouter.post("/tasks", requireEditor, async (req, res, next) => {
       `insert into task (org_id, editor_id, channel_id, title, description,
                          status, priority, due_date, recurrence, task_type, content_format,
                          budget_hours, accepted, budget_started_at, completed_at)
-       values ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13, case when $12 is not null and $13 then now() else null end,
+       values ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13, case when $12::numeric is not null and $13 then now() else null end,
                case when $6 = 'done' then now() else null end)
        returning id`,
       [
@@ -343,7 +343,7 @@ async function spawnNextOccurrence(taskId, recurrence) {
     const budgetHours = await resolveBudgetHours(t.org_id, t.editor_id, t.content_format);
     const { rows } = await pool.query(
       `insert into task (org_id, editor_id, channel_id, title, description, priority, due_date, recurrence, status, task_type, content_format, budget_hours, budget_started_at)
-       values ($1,$2,$3,$4,$5,$6, ${nextDue}, $8, 'todo', $9, $10, $11, case when $11 is not null then now() else null end) returning id`,
+       values ($1,$2,$3,$4,$5,$6, ${nextDue}, $8, 'todo', $9, $10, $11, case when $11::numeric is not null then now() else null end) returning id`,
       [t.org_id, t.editor_id, t.channel_id, t.title, t.description, t.priority, t.due_date, recurrence, t.task_type, t.content_format, budgetHours],
     );
     const newId = rows[0].id;
